# docs/swagger/libraryapi_swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Webtor API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/library": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "The torrents saved to the account, newest first by default.\n\n`type` filters the same way the library tabs do: `movies` and `series` mean a torrent with at least\none recognized film or show, which is decided by enrichment — a torrent added a moment ago may not\nbe classified yet and shows up only under `all`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "List your library",
                "parameters": [
                    {
                        "enum": [
                            "all",
                            "movies",
                            "series"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "Filter",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "recent",
                            "name"
                        ],
                        "type": "string",
                        "default": "recent",
                        "description": "Sort order",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 100, max 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/libapi.LibraryListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Takes a resource id that is already in the store — `POST /resource` puts it there and returns one.\n\nIdempotent: adding a resource already in the library answers `200` with the existing entry, and only\na real insert answers `201`. Adding also queues\nenrichment (title, year, poster, movie/series classification), which runs in the background — the\nentry is usable immediately but may not answer to `type=movies` for a little while.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "Add a torrent to your library",
                "parameters": [
                    {
                        "description": "Resource to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/libapi.LibraryAddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Already in the library",
                        "schema": {
                            "$ref": "#/definitions/libapi.LibraryItem"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/libapi.LibraryItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "The resource is not in the store",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/library/{resource_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Answers `404` when the resource is not in your library, which is how a client checks membership\nbefore offering to add or remove it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "Get one library entry",
                "parameters": [
                    {
                        "type": "string",
                        "example": "08ada5a7a6183aae1e09d831df6748d566095a10",
                        "description": "Infohash",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/libapi.LibraryItem"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes the entry from your library. The resource itself stays in the store and keeps working —\nthis is not a takedown, and a Vault pledge on it is a separate thing (`DELETE /vault/pledges`).\n\nNot idempotent: removing something that is not there answers `404`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "Remove a torrent from your library",
                "parameters": [
                    {
                        "type": "string",
                        "example": "08ada5a7a6183aae1e09d831df6748d566095a10",
                        "description": "Infohash",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Changes the name the item is shown under everywhere — the library, WebDAV and S3 all read this\nrow. It does not touch the torrent itself.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "library"
                ],
                "summary": "Rename a library entry",
                "parameters": [
                    {
                        "type": "string",
                        "example": "08ada5a7a6183aae1e09d831df6748d566095a10",
                        "description": "Infohash",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/libapi.LibraryRenameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/libapi.LibraryItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Who the key belongs to, which plan the account is on, what the key is allowed to do, and the\npreferences that affect what the other endpoints return.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Account state",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/libapi.ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partial update: a field left out of the body keeps its current value. Requires the `api:write`\nscope.\n\nPreferences are read on the hot path and cached for a few minutes, so a change can take that long\nto show up in poster URLs — this endpoint invalidates its own cache entry, but other replicas\nexpire on their own schedule.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Update account preferences",
                "parameters": [
                    {
                        "description": "Fields to change",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/libapi.ProfileSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/libapi.ProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/resource": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Takes a `.torrent` file or a magnet URI in the request body and puts it in the store, returning the\nresource id (its infohash) that every other endpoint takes. A magnet is resolved against the\nBitTorrent network first, which can take minutes for a torrent nobody is seeding.\n\nStoring does **not** add the torrent to your library — `POST /library` does, and takes the id this\nreturns.\n\nIdentical in shape to rest-api's `POST /resource/`.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource"
                ],
                "summary": "Store a resource",
                "parameters": [
                    {
                        "example": "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10",
                        "description": "Raw .torrent bytes, or a magnet URI",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ResourceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "The torrent could not be fetched",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "A magnet took too long to resolve",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/resource/{resource_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Name, size, file count and magnet URI of a stored resource. Single-file torrents also carry the\nfile itself, so the common case needs no `/list` round-trip.\n\nAppending `.torrent` to the id returns the torrent file instead\n(`GET /resource/{resource_id}.torrent`, `application/x-bittorrent`) — same convention as rest-api.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource"
                ],
                "summary": "Get a resource",
                "parameters": [
                    {
                        "type": "string",
                        "example": "08ada5a7a6183aae1e09d831df6748d566095a10",
                        "description": "Infohash, optionally suffixed with .torrent",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ResourceResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests with this key — the `Retry-After` header says how long to wait",
                        "schema": {
                            "$ref": "#/definitions/libapi.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "libapi.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/libapi.Error"
                }
            }
        },
        "libapi.LibraryAddRequest": {
            "type": "object",
            "required": [
                "resource_id"
            ],
            "properties": {
                "resource_id": {
                    "type": "string",
                    "example": "08ada5a7a6183aae1e09d831df6748d566095a10"
                }
            }
        },
...
        "libapi.LibraryListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/libapi.LibraryItem"
                    }
                },
                "items_count": {
                    "description": "Count is the total number of matching items, not the size of this page.",
                    "type": "integer",
                    "example": 42
                },
                "limit": {
                    "type": "integer",
                    "example": 100
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "sort": {
                    "type": "string",
                    "example": "recent"
                },
                "type": {
                    "type": "string",
                    "example": "all"
                }
            }
        },
        "libapi.LibraryRenameRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Sintel (2010)"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "API key from your profile page, as `Bearer \u003ckey\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}